requirements.txt
altgraph==0.17.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmudict==1.0.13
colorama==0.4.6
grammarbot==0.2.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.13.0
joblib==1.3.2
language-tool-python==2.7.1
mypy-extensions==1.0.0
nltk==3.8.1
packaging==23.1
pathspec==0.11.2
pefile==2023.2.7
platformdirs==3.10.0
py-readability-metrics==1.4.5
pyenchant==3.2.2
pyinstaller-hooks-contrib==2023.8
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pywin32-ctypes==0.2.2
regex==2023.8.8
requests==2.31.0
textblob==0.17.1
tqdm==4.66.1
urllib3==2.0.4
zipp==3.17.0
